Implementation of decryption functions

+++ BaosMini/include/utility/DataReader.hpp
@@ -2,11 +2,12 @@
 #define DATA_READER_HPP
 
 #include <Windows.h>
+#include <string>
 #include "ChecksumCalculator.hpp"
 
 namespace DataReader
 {
-	size_t recieveTelegram(HANDLE serialHandle, unsigned char *telegramCharArray);
+	unsigned int recieveTelegram(HANDLE serialHandle, unsigned char *telegramCharArray);
 }
 
 #endif // DATA_READER_HPP
